requirements.txt
torch
numpy
ipdb
wandb
quinine
pyyaml